+++ wifi/wifi_copy.h
@@ -1,0 +1,26 @@
+/**
+ * @file wifi.h
+ * @author Martin Garncarz (246815@vutbr.cz)
+ * @brief Main functionality of the second (wifi) core
+ * @version 0.1
+ * @date 2024-07-09
+ * 
+ * @copyright Copyright (c) 2024
+ * 
+ */
+
+#ifndef __WIFI_COPY_H__
+#define __WIFI_COPY_H__
+
+#include "common/structs.h"
+
+
+/**
+ * @brief Wifi & network entry point
+ * 
+ */
+extern void wifi_main();
+
+extern bool wifi_is_running(void);
+
+#endif